+++ .gitignore
@@ -2,8 +2,8 @@
 /bin
 /obj
 /src/.vs
-/src/bin
-/src/obj
-/src/settings.json
-/src/Properties
+/src/**/bin
+/src/**/obj
+/src/**/settings.json
+/src/**/Properties
 /src/Gps2Yandex.csproj.user
