--- src/ApiYandex/Services/Config.cs
@@ -1,9 +1,0 @@
-﻿namespace Gps2Yandex.ApiYandex.Services
-{
-    internal class Config
-    {
-        public string Clid { get; set; }
-        public string Host { get; set; }
-        public int Interval { get; set; }
-    }
-}
