--- src/Dataset/Services/Config.cs
@@ -1,7 +1,0 @@
-﻿namespace Gps2Yandex.Dataset.Services
-{
-    internal class Config
-    {
-        public string Directory { get; set; }
-    }
-}
