+++ src/Gps2Yandex.Model/Services/Config.cs
@@ -1,0 +1,7 @@
+﻿namespace Gps2Yandex.Model.Services
+{
+    internal class Config
+    {
+        public string Directory { get; set; }
+    }
+}
