+++ src/Gps2Yandex.Wialon/Models/PointGPS.cs
