+++ src/Gps2Yandex.Wialon/Services/Config.cs
