+++ src/Gps2Yandex.Yandex/Services/Config.cs
@@ -1,0 +1,9 @@
+﻿namespace Gps2Yandex.Yandex.Services
+{
+    internal class Config
+    {
+        public string Clid { get; set; }
+        public string Host { get; set; }
+        public int Interval { get; set; }
+    }
+}
