--- src/Gps2Yandex.csproj
@@ -1,33 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <AssemblyName>Gps2Yandex</AssemblyName>
-    <RootNamespace>Gps2Yandex</RootNamespace>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="3.1.4" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.FileExtensions" Version="3.1.4" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Json" Version="3.1.4" />
-    <PackageReference Include="Microsoft.Extensions.Hosting" Version="3.1.4" />
-  </ItemGroup>
-
-  <Target Name="FileCopyAfterBuild" AfterTargets="AfterBuild">
-    <Copy SourceFiles="settings.json.sample" DestinationFiles="$(OutDir)settings.json" Condition="!Exists('$(OutDir)settings.json')" />
-    <MakeDir Directories="$(OutDir)source" Condition="!Exists('$(OutDir)source')" />
-    <Copy SourceFiles="SourceSample\route.txt.sample" DestinationFiles="$(OutDir)source\route.txt" Condition="!Exists('$(OutDir)source\route.txt')" />
-    <Copy SourceFiles="SourceSample\schedule.txt.sample" DestinationFiles="$(OutDir)source\schedule.txt" Condition="!Exists('$(OutDir)source\schedule.txt')" />
-    <Copy SourceFiles="SourceSample\transport.txt.sample" DestinationFiles="$(OutDir)source\transport.txt" Condition="!Exists('$(OutDir)source\transport.txt')" />
-  </Target>
-
-  <Target Name="SettingsFileCopyPublish" AfterTargets="CopyFilesToPublishDirectory">
-    <Copy SourceFiles="settings.json.sample" DestinationFiles="$(PublishDir)settings.json" />
-    <MakeDir Directories="$(PublishDir)source" />
-    <Copy SourceFiles="SourceSample\route.txt.sample" DestinationFiles="$(PublishDir)source\route.txt" />
-    <Copy SourceFiles="SourceSample\schedule.txt.sample" DestinationFiles="$(PublishDir)source\schedule.txt" />
-    <Copy SourceFiles="SourceSample\transport.txt.sample" DestinationFiles="$(PublishDir)source\transport.txt" />
-  </Target>
-
-</Project>
