+++ src/Gps2Yandex/SourceSample/route.txt.sample
