+++ src/Gps2Yandex/SourceSample/schedule.txt.sample
