+++ src/Gps2Yandex/settings.json.sample
