--- src/Wialon/Extensions/ServiceCollectionExtensions.cs
@@ -1,14 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-
-using Gps2Yandex.Wialon.Services;
-
-namespace Gps2Yandex.Wialon.Extensions
-{
-    public static class ServiceCollectionExtensions
-    {
-        public static IServiceCollection AddWialonListener(this IServiceCollection serviceCollection)
-        {
-            return serviceCollection.AddHostedService<ListenerRetransmission>();
-        }
-    }
-}
